chore: code cleanups
