+++ src/engineV2/engine.ts
@@ -1,14 +1,11 @@
 import { Mutex } from "async-mutex";
-import { ScaleLinear, scaleLinear } from "d3-scale";
 import { produce } from "immer";
 import merge from "merge-k-sorted-arrays";
-import hash from "object-hash";
 import BTree from 'sorted-btree';
-import { newBatchDoneMessage, newJobUpdatedMessage } from "src/plugins_engine/websocket";
+import { newBatchDoneMessage, newJobUpdatedMessage } from "src/plugins_engine/messages";
 import { v4 as uuidv4 } from 'uuid';
-import { Adapter, Param, PluginRef, QueryProvider } from "~lib/adapters";
+import { Adapter, PluginRef } from "~lib/adapters";
 import { asDateField, asDisplayString, compareProcessedData, ProcessedData } from "~lib/adapters/logTypes";
-import { FullDate } from "~lib/dateUtils";
 import { DisplayResults, Events } from "~lib/displayTypes";
 import { ResponseHandler } from "~lib/networkTypes";
 import { processEval } from "~lib/pipelineEngine/eval";
@@ -20,129 +17,10 @@ import { processTable } from "~lib/pipelineEngine/table";
 import { processTimeChart } from "~lib/pipelineEngine/timechart";
 import { processWhere } from "~lib/pipelineEngine/where";
 import { parse, ParsedQuery, PipelineItem } from "~lib/qql";
-import { ControllerIndexParam, Search } from "~lib/qql/grammar";
-import { createSignal, measureTime, Signal } from "~lib/utils";
-
-export type TaskRef = string & { _tr: never }; // A unique identifier for a plugin
-export type QueryTask = {
-    id: TaskRef;
-    input: QueryInput;
-    status: "running" | "completed" | "failed" | "canceled";
-    createdAt: Date;
-}
-
-export type QueryInput = {
-    searchTerm: string;
-    queryOptions: SerializeableParams;
-}
-
-export type QueryTaskState = {
-    task: QueryTask;
-
-    subTasks: SubTask[];
-    abortController: AbortController;
-    index: BTree<number, ProcessedData[]>
-    displayResults: DisplayResults;
-    finishedQuerying: Signal;
-    mutex: Mutex;
-}
-
-export type QueryExecutionHistory = {
-    params: ControllerIndexParam[];
-    search: Search;
-    start: FullDate;
-    end: FullDate;
-    instanceRef: InstanceRef;
-};
-
-
-export type InstanceRef = string & { _ir: never }; // A unique identifier for a plugin instance
-export type SearchProfileRef = string & { _spr: never }; // A unique identifier for a search profile
-export type SearchProfile = {
-    name: SearchProfileRef;
-    instances: InstanceRef[];
-}
-
-// MUST BE SERIALIZABLE
-export type PluginInstance = {
-    id: string;
-    name: InstanceRef;
-    description: string;
-    pluginRef: PluginRef;
-}
-
-// MUST BE SERIALIZABLE
-export type SerializableAdapter = {
-    ref: PluginRef;
-    name: string;
-    description: string;
-    version: string;
-    params: Param[];
-}
-
-// MUST BE SERIALIZABLE
-export type SerializeableParams = {
-    fromTime: Date,
-    toTime: Date,
-    limit: number,
-    isForced: boolean,
-}
-
-
-export type PluginInstanceContainer = {
-    instance: PluginInstance;
-    provider: QueryProvider;
-}
-
-export type MessageSender = {
-    batchDone: (jobId: string, data: ProcessedData[]) => void;
-    jobUpdated: (job: QueryTask) => void;
-    urlNavigate: (url: string) => void;
-}
-
-export type PageResponse<T> = {
-    data: T[];
-    total: number;
-    limit: number;
-    next: number | null; // Reference to the next page, if any
-    prev: number | null; // Reference to the previous page, if any
-}
-
-export type TableDataResponse = PageResponse<ProcessedData> & {
-    columns: string[]; // List of column names in the table
-    columnLengths: Record<string, number>; // Length of each column for display purposes
-}
-
-export type ClosestPoint = {
-    closest: number | null
-    index: number | null // Index of the closest point in the data array
-}
-
-export type SubTaskRef = string & { _: never }; // A unique identifier for a subtask
-export type SubTask = {
-    id: SubTaskRef;
-    createdAt: Date;
-    instanceRef: InstanceRef;
-    cacheKey: string;
-    isReady: Promise<void>;
-}
-
-export type JobBatchFinished = {
-    scale: {
-        from: number;
-        to: number;
-    }
-    views: {
-        events: {
-            total: number;
-            buckets: { timestamp: number, count: number }[];
-            autoCompleteKeys: string[];
-        },
-        table?: {
-            totalRows: number;
-        }
-    }
-}
+import { createSignal, measureTime } from "~lib/utils";
+import { CacheRecord, QueryCacheHolder } from "./cache";
+import { ClosestPoint, InstanceRef, JobBatchFinished, PageResponse, PluginInstance, PluginInstanceContainer, QueryExecutionHistory, QueryTask, QueryTaskState, SearchProfile, SearchProfileRef, SerializableAdapter, SerializeableParams, SubTask, SubTaskRef, TableDataResponse, TaskRef } from "./types";
+import { calculateBuckets, getScale } from "./utils";
 
 export class Engine {
     private supportedPlugins: Adapter[] = [];
@@ -659,169 +537,4 @@ export class Engine {
         console.log(`Query task ${taskId} cancelled`);
         this.messageSender.sendMessage(newJobUpdatedMessage(taskId, taskState.task.status)); // Notify the client about the cancellation
     }
-}
-
-
-type CacheRecord = {
-    id: string; // Unique identifier for the cache record
-    referencingTasks: Set<TaskRef>;
-    data: ProcessedData[];
-    lastAccessed: Date;
-    key: string;
-    identifier: QueryExecutionHistory;
-    status: "running" | "completed" | "failed" | "canceled";
-    promise: Promise<void>;
-}
-
-class QueryCacheHolder {
-    private cache: Record<string, CacheRecord> = {};
-    private cacheMutex = new Mutex();
-
-    public inCache(
-        identifier: QueryExecutionHistory
-    ): Promise<boolean> {
-        return this.cacheMutex.runExclusive(() => {
-            const key = hash(identifier);
-            return Promise.resolve(!!this.cache[key]);
-        });
-    }
-
-    public forceRemoveFromCache(
-        identifier: QueryExecutionHistory
-    ): Promise<void> {
-        return this.cacheMutex.runExclusive(() => {
-            const key = hash(identifier);
-            if (this.cache[key]) {
-                delete this.cache[key];
-                console.log(`Cache record removed for key: ${key}`);
-            } else {
-                console.warn(`No cache record found for key: ${key}`);
-            }
-        });
-    }
-
-    public addToCache(
-        identifier: QueryExecutionHistory,
-        taskId: TaskRef,
-        promiseFactory: (cacheRecord: CacheRecord) => Promise<void>,
-    ): Promise<CacheRecord> {
-        return this.cacheMutex.runExclusive(() => {
-            const key = hash(identifier);
-            if (this.cache[key]) {
-                throw new Error(`Cache already exists for key: ${key}. Use getFromCacheByKey to retrieve it.`);
-            }
-
-            const newRecord = {
-                id: uuidv4(), // Generate a unique ID for the cache record
-                key: key,
-                referencingTasks: new Set([taskId]),
-                data: [],
-                status: "running", // Initial status can be set to running
-                lastAccessed: new Date(),
-                identifier: identifier,
-                promise: Promise.resolve(), // Initialize with a resolved promise
-            } as CacheRecord;
-            newRecord.promise = promiseFactory(newRecord);
-
-            this.cache[key] = newRecord;
-            console.log(`Cache record added for key: ${key}`);
-
-            return this.cache[key];
-        });
-    }
-
-    public referenceCache(
-        identifier: QueryExecutionHistory,
-        taskId: TaskRef
-    ): Promise<CacheRecord> {
-        return this.cacheMutex.runExclusive(() => {
-            const key = hash(identifier);
-            if (!this.cache[key]) {
-                throw new Error(`Cache miss for key: ${key}`);
-            }
-
-            // Add the taskId to the referencing tasks
-            this.cache[key].referencingTasks.add(taskId);
-            this.cache[key].lastAccessed = new Date(); // Update last accessed time
-            return Promise.resolve(this.cache[key]);
-        });
-    }
-
-    public getFromCacheByKey(
-        key: string
-    ): Promise<CacheRecord> {
-        return this.cacheMutex.runExclusive(() => {
-            const record = this.cache[key];
-            if (!record) {
-                throw new Error(`Cache miss for key: ${key}`);
-            }
-
-            // Update last accessed time
-            record.lastAccessed = new Date();
-            return Promise.resolve(record);
-        });
-    }
-
-    public getFromCache(
-        identifier: QueryExecutionHistory
-    ): Promise<CacheRecord> {
-        const key = hash(identifier);
-        return this.getFromCacheByKey(key);
-    }
-
-    public removeFromCacheByKey(
-        key: string,
-        taskId: TaskRef
-    ): Promise<void> {
-        return this.cacheMutex.runExclusive(() => {
-            if (this.cache[key]) {
-                this.cache[key].referencingTasks.delete(taskId);
-                if (this.cache[key].referencingTasks.size === 0) {
-                    delete this.cache[key];
-                }
-            }
-        });
-    }
-}
-
-const getScale = (selectedStartTime: Date, selectedEndTime: Date) => {
-    if (!selectedStartTime || !selectedEndTime) {
-        return;
-    }
-
-    return scaleLinear().domain([
-        selectedStartTime.getTime(),
-        selectedEndTime.getTime(),
-    ]);
-}
-
-
-const calculateBuckets = (scale: ScaleLinear<number, number, unknown> | undefined, data: ProcessedData[]) => {
-    if (!scale) {
-        return [];
-    }
-
-    const buckets: Record<number, number> = {};
-    const ticks = scale.ticks(100);
-
-    data.forEach((object) => {
-        // round timestamp to the nearest tick
-        const timestamp = ticks.reduce((prev, curr) => {
-            const thisTimestamp = asDateField(object.object._time).value;
-
-            return Math.abs(curr - thisTimestamp) < Math.abs(prev - thisTimestamp)
-                ? curr
-                : prev;
-        });
-        if (!buckets[timestamp]) {
-            buckets[timestamp] = 0;
-        }
-
-        buckets[timestamp] += 1;
-    });
-
-    return Object.entries(buckets).map(([timestamp, count]) => ({
-        timestamp: parseInt(timestamp),
-        count,
-    }));
 }
