+++ src/renderer/core/common/DefaultQueryProvider.ts
@@ -1,7 +1,7 @@
-import { ClosestPoint, PageResponse, TableDataResponse, TaskRef } from "src/engineV2/engine";
 import { ProcessedData } from "~lib/adapters/logTypes";
 import { DisplayResults } from "~lib/displayTypes";
 import { AwaitableTask, QueryOptions } from "./interface";
+import { ClosestPoint, PageResponse, TableDataResponse, TaskRef } from "src/engineV2/types";
 
 
 export class DefaultQueryProvider {
