+++ src/renderer/core/store/queryState.ts
@@ -1,7 +1,7 @@
 import { scaleLinear } from 'd3-scale';
 import { atom, createStore } from 'jotai';
 import React from 'react';
-import { JobBatchFinished } from 'src/engineV2/engine';
+import { JobBatchFinished } from 'src/engineV2/types';
 import { DisplayResults, Events } from '~lib/displayTypes';
 import { allData } from '~lib/qql';
